+++ backend/src/main/java/kr/co/nogibackend/domain/notion/NotionClient.java
@@ -7,21 +7,21 @@ import kr.co.nogibackend.domain.notion.dto.info.NotionDatabaseInfo;
 import kr.co.nogibackend.domain.notion.dto.info.NotionGetAccessInfo;
 import kr.co.nogibackend.domain.notion.dto.info.NotionInfo;
 import kr.co.nogibackend.domain.notion.dto.info.NotionPageInfo;
-import kr.co.nogibackend.domain.notion.dto.request.NotionGetAccessTokenRequest;
+import kr.co.nogibackend.interfaces.notion.dto.request.NotionGetAccessTokenRequest;
 
 public interface NotionClient {
 
-  NotionInfo<NotionPageInfo> getPagesFromDatabase(String BotToken, String databaseId,
-      Map<String, Object> request);
+	NotionInfo<NotionPageInfo> getPagesFromDatabase(String BotToken, String databaseId,
+			Map<String, Object> request);
 
-  NotionInfo<NotionBlockInfo> getBlocksFromPage(String BotToken, String pageId, String startCursor);
+	NotionInfo<NotionBlockInfo> getBlocksFromPage(String BotToken, String pageId, String startCursor);
 
-  byte[] getBlockImage(URI baseUri);
+	byte[] getBlockImage(URI baseUri);
 
-  NotionPageInfo updatePageStatus
-      (String BotToken, String pageId, Map<String, Object> request);
+	NotionPageInfo updatePageStatus
+			(String BotToken, String pageId, Map<String, Object> request);
 
-  NotionDatabaseInfo getDatabase(String BotToken, String databaseId);
+	NotionDatabaseInfo getDatabase(String BotToken, String databaseId);
 
-  NotionGetAccessInfo getAccessToken(String basicToken, NotionGetAccessTokenRequest request);
+	NotionGetAccessInfo getAccessToken(String basicToken, NotionGetAccessTokenRequest request);
 }
