--- backend/src/main/java/kr/co/nogibackend/domain/notion/constant/NotionPropertyValue.java
@@ -1,19 +1,0 @@
-package kr.co.nogibackend.domain.notion.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum NotionPropertyValue {
-
-  STATUS_DRAFTING("작성중"),
-  STATUS_COMPLETED("작성완료"),
-  STATUS_FAILED("실패"),
-  STATUS_DONE("완료");
-
-  private final String name;
-
-  NotionPropertyValue(String name) {
-    this.name = name;
-  }
-
-}
