+++ backend/src/main/java/kr/co/nogibackend/domain/notion/dto/constant/NotionPropertyKey.java
@@ -1,0 +1,19 @@
+package kr.co.nogibackend.domain.notion.dto.constant;
+
+import lombok.Getter;
+
+@Getter
+public enum NotionPropertyKey {
+
+	STATUS("nogiStatus"),
+	CATEGORY("nogiCategory"),
+	TITLE("nogiTitle"),
+	CREATE_DATE("nogiCommitDate");
+
+	private final String name;
+
+	NotionPropertyKey(String name) {
+		this.name = name;
+	}
+
+}
