--- backend/src/main/java/kr/co/nogibackend/domain/notion/dto/response/NotionConnectionResponse.java
@@ -1,8 +1,0 @@
-package kr.co.nogibackend.domain.notion.dto.response;
-
-public record NotionConnectionResponse(
-    boolean isConnection,
-    String message
-) {
-
-}
