+++ backend/src/main/java/kr/co/nogibackend/interfaces/notion/dto/response/NotionConnectionResponse.java
@@ -1,0 +1,8 @@
+package kr.co.nogibackend.interfaces.notion.dto.response;
+
+public record NotionConnectionResponse(
+		boolean isConnection,
+		String message
+) {
+
+}
