+++ backend/src/test/java/kr/co/nogibackend/domain/notion/markdown/NotionConvertMarkdownTable.java
@@ -3,10 +3,10 @@ package kr.co.nogibackend.domain.notion.markdown;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
-import kr.co.nogibackend.domain.notion.NotionService;
 import kr.co.nogibackend.domain.notion.dto.content.NotionTableRowCellsContent;
 import kr.co.nogibackend.domain.notion.dto.content.NotionTableRowContent;
 import kr.co.nogibackend.domain.notion.dto.info.NotionBlockInfo;
+import kr.co.nogibackend.domain.notion.service.NotionService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
